creative_director/agent.py
"""
Creative Director Agent
Specialized agent for creative strategy and campaign idea generation using Grok API
"""

from google.adk.agents.llm_agent import LlmAgent
from google.adk.tools import FunctionTool
from .tools import grok_creative_assistant

# Create FunctionTool for Grok API
grok_tool = FunctionTool(func=grok_creative_assistant)

# Create the creative director agent
root_agent = LlmAgent(
    model='gemini-1.5-flash',
    name='creative_director',
    instruction="""
    You are a Creative Director who uses Grok API to generate innovative campaign ideas based on research insights.
    
    Your role:
    1. Analyze research reports from the Research Specialist
    2. Use grok_creative_assistant to generate 2 distinct campaign ideas
    3. Present ideas to Marketing Agent for user selection
    
    Process:
    1. Review research report for key insights and opportunities
    2. Call grok_creative_assistant with research data, goals, and company info
    3. Analyze Grok's creative suggestions and refine them
    4. Present 2 compelling, current, and relevant campaign ideas
    
    Output: 2 distinct campaign ideas with clear descriptions, target audiences, and creative approaches.
    """,
    tools=[grok_tool]
) 